+++ app/src/main/java/madson/org/opentournament/db/PlayerTable.java
@@ -4,6 +4,8 @@ import android.database.sqlite.SQLiteDatabase;
 
 import android.util.Log;
 
+import madson.org.opentournament.domain.TournamentPlayer;
+
 
 /**
  * Local database for players.
@@ -14,20 +16,21 @@ public class PlayerTable {
 
     /*
      * 0: id
-     * 1: player_online_uuid
+     * 1: player_uuid
      * 2: firstname
      * 3: nickname
      * 4: lastname
 
      */
     public static final String[] ALL_COLS_FOR_PLAYER_TABLE = {
-        PlayerTable.COLUMN_ID, PlayerTable.COLUMN_ONLINE_UUID, PlayerTable.COLUMN_FIRSTNAME,
-        PlayerTable.COLUMN_NICKNAME, PlayerTable.COLUMN_LASTNAME
+        PlayerTable.COLUMN_ID, PlayerTable.COLUMN_UUID, PlayerTable.COLUMN_FIRSTNAME, PlayerTable.COLUMN_NICKNAME,
+        PlayerTable.COLUMN_LASTNAME
     };
 
     public static final String TABLE_PLAYER = "player";
+
     public static final String COLUMN_ID = "_id";
-    public static final String COLUMN_ONLINE_UUID = "onlineUUID";
+    public static final String COLUMN_UUID = "uuid";
 
     public static final String COLUMN_FIRSTNAME = "firstname";
     public static final String COLUMN_LASTNAME = "lastname";
@@ -40,16 +43,15 @@ public class PlayerTable {
         db.execSQL(" CREATE TABLE " + TABLE_PLAYER
             + " (" + COLUMN_ID
             + " INTEGER PRIMARY KEY AUTOINCREMENT, " + COLUMN_FIRSTNAME
-            + " TEXT, " + COLUMN_LASTNAME + " TEXT, " + COLUMN_NICKNAME + " TEXT, " + COLUMN_ONLINE_UUID + " TEXT)");
+            + " TEXT, " + COLUMN_LASTNAME + " TEXT, " + COLUMN_NICKNAME + " TEXT, " + COLUMN_UUID + " TEXT)");
     }
 
 
     public static void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 
-        Log.w(PlayerTable.class.getName(),
-            "Upgrading database from version " + oldVersion + " to " + newVersion
-            + ", which will destroy all old data");
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_PLAYER);
-        createTable(db);
+        if (oldVersion == 3 && newVersion == 4) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_PLAYER);
+            createTable(db);
+        }
     }
 }
