+++ app/src/main/java/madson/org/opentournament/db/TournamentPlayerTable.java
@@ -15,32 +15,31 @@ public class TournamentPlayerTable {
     /*
      * 0: id
      * 1: tournament_id
-     * 2: player_id
-     * 3: player_online_uuid
-     * 4: firstname
-     * 5: nickname
-     * 6: lastname
-     * 7: teamname
-     * 8: faction
-     * 9: meta
-     * 10: dummy
-     * 11: dropped_in_round
+     * 2: player_uuid
+     * 3 firstname
+     * 4: nickname
+     * 5: lastname
+     * 6: teamname
+     * 7: faction
+     * 8: meta
+     * 9: dummy
+     * 10: dropped_in_round
+     * 11: local
      */
     public static final String[] ALL_COLS_FOR_TOURNAMENT_PLAYER_TABLE = {
         TournamentPlayerTable.COLUMN_ID, TournamentPlayerTable.COLUMN_TOURNAMENT_ID,
-        TournamentPlayerTable.COLUMN_PLAYER_ID, TournamentPlayerTable.COLUMN_PLAYER_ONLINE_UUID,
-        TournamentPlayerTable.COLUMN_FIRSTNAME, TournamentPlayerTable.COLUMN_NICKNAME,
-        TournamentPlayerTable.COLUMN_LASTNAME, TournamentPlayerTable.COLUMN_TEAMNAME,
-        TournamentPlayerTable.COLUMN_FACTION, TournamentPlayerTable.COLUMN_META, TournamentPlayerTable.COLUMN_DUMMY,
-        TournamentPlayerTable.COLUMN_DROPPED_IN_ROUND
+        TournamentPlayerTable.COLUMN_PLAYER_UUID, TournamentPlayerTable.COLUMN_FIRSTNAME,
+        TournamentPlayerTable.COLUMN_NICKNAME, TournamentPlayerTable.COLUMN_LASTNAME,
+        TournamentPlayerTable.COLUMN_TEAMNAME, TournamentPlayerTable.COLUMN_FACTION, TournamentPlayerTable.COLUMN_META,
+        TournamentPlayerTable.COLUMN_DUMMY, TournamentPlayerTable.COLUMN_DROPPED_IN_ROUND,
+        TournamentPlayerTable.COLUMN_LOCAL
     };
 
     public static final String TABLE_TOURNAMENT_PLAYER = "tournament_player";
 
     public static final String COLUMN_ID = "_id";
     public static final String COLUMN_TOURNAMENT_ID = "tournament_id";
-    public static final String COLUMN_PLAYER_ID = "player_id";
-    public static final String COLUMN_PLAYER_ONLINE_UUID = "player_online_uuid";
+    public static final String COLUMN_PLAYER_UUID = "player_uuid";
 
     public static final String COLUMN_FIRSTNAME = "firstname";
     public static final String COLUMN_NICKNAME = "nickname";
@@ -54,17 +53,18 @@ public class TournamentPlayerTable {
     public static final String COLUMN_DUMMY = "dummy";
 
     public static final String COLUMN_DROPPED_IN_ROUND = "dropped_on_round";
+    public static final String COLUMN_LOCAL = "local";
 
     public static void createTable(SQLiteDatabase db) {
 
         Log.i(TournamentPlayerTable.class.getName(), "create tournament_player table");
 
         db.execSQL(" CREATE TABLE " + TABLE_TOURNAMENT_PLAYER
-            + " (" + COLUMN_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " + COLUMN_TOURNAMENT_ID + " INTEGER,"
-            + COLUMN_PLAYER_ID + " INTEGER, " + COLUMN_PLAYER_ONLINE_UUID + " TEXT," + COLUMN_FIRSTNAME
-            + " TEXT, " + COLUMN_NICKNAME + " TEXT, " + COLUMN_LASTNAME + " TEXT, " + COLUMN_TEAMNAME + " TEXT, "
-            + COLUMN_FACTION + " TEXT, " + COLUMN_META + " TEXT, " + COLUMN_DUMMY + " INTEGER, "
-            + COLUMN_DROPPED_IN_ROUND + " INTEGER)");
+            + " (" + COLUMN_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " + COLUMN_TOURNAMENT_ID + " TEXT,"
+            + COLUMN_PLAYER_UUID + " TEXT," + COLUMN_FIRSTNAME + " TEXT, " + COLUMN_NICKNAME + " TEXT, "
+            + COLUMN_LASTNAME + " TEXT, " + COLUMN_TEAMNAME + " TEXT, " + COLUMN_FACTION + " TEXT, " + COLUMN_META
+            + " TEXT, " + COLUMN_DUMMY + " INTEGER, " + COLUMN_DROPPED_IN_ROUND + " INTEGER, " + COLUMN_LOCAL
+            + " INTEGER)");
     }
 
 
@@ -73,7 +73,10 @@ public class TournamentPlayerTable {
         Log.w(TournamentPlayerTable.class.getName(),
             "Upgrading database from version " + oldVersion + " to " + newVersion
             + ", which will destroy all old data");
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_TOURNAMENT_PLAYER);
-        createTable(db);
+
+        if (oldVersion == 3 && newVersion == 4) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_TOURNAMENT_PLAYER);
+            createTable(db);
+        }
     }
 }
